src/examples/data_loading/requirements.txt
datasets>=2.0.0
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0



